% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_method.R
\name{print.fit_model}
\alias{print.fit_model}
\title{Print method for fit_model objects}
\usage{
print.fit_model(x, ...)
}
\arguments{
\item{x}{A fit_model object returned by fit_model()}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the input object
}
\description{
Provides concise display of fit_model objects
}
