% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_GP}
\alias{test_GP}
\title{Test dataset for Gaussian Process model}
\format{
A data frame with 40 rows and 7 columns:
\describe{
  \item{X}{X spatial coordinate (1 to 100)}
  \item{Y}{Y spatial coordinate (1 to 100)}
  \item{X1}{Transformed covariate: sqrt(|Y - 60|)}
  \item{X2}{Transformed covariate: sqrt(|X - 20|)}
  \item{Z}{True (uncensored) response variable}
  \item{censored}{Censored response variable}
  \item{is_censored}{Binary indicator: 1 if censored, 0 otherwise}
}
}
\source{
Simulated data for package demonstration
}
\usage{
test_GP
}
\description{
Hold-out test set corresponding to \code{\link{grille_GP}}, containing
20% of the original simulated data for model validation.
}
\details{
This is the test split from the same simulation that produced \code{\link{grille_GP}}.
Use this dataset to evaluate prediction performance.
}
\examples{
# Use with grille_GP for prediction
data(grille_GP)
data(test_GP)

# Fit model on training data
fit_gp <- fit_model(
  Y_obs = grille_GP$censored,
  locs_obs = as.matrix(grille_GP[, c("X", "Y")]),
  X_obs = as.matrix(grille_GP[, c("X1", "X2")]),
  censored_indices = rep(0, nrow(grille_GP)),
  M = 30,
  bayesian = FALSE
)

# Make predictions on test set (when predict function is available)
# predictions <- predict(fit_gp,
#                        newlocs = as.matrix(test_GP[, c("X", "Y")]),
#                        newX = as.matrix(test_GP[, c("X1", "X2")]))

}
\seealso{
\code{\link{grille_GP}}
}
\keyword{datasets}
